Remove unnecessary file.
